preset: move params to the right file and reorder functions and types
